requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncssh>=2.14.0
pyyaml>=6.0
pydantic>=2.0
